Make Butler an abstract class, with DirectButler as default implementation (DM-41116)
`Butler` class has been renamed to `DirectButler`, and a new intermediate abstract
base class `Butler` is introduced to define full butler interface (on top of
`LimitedButler`). Butlers are now instantiated using a factory method
`Butler.from_config`, call to `Butler` constructor will also return an instance
of a concrete butler class, just as `Butler.from_config`, but `Butler(...)`
will cause `mypy` complaints about an abstract class being instantiated.
