+++ python/lsst/daf/butler/script/queryDatasetTypes.py
@@ -61,7 +61,7 @@ def queryDatasetTypes(repo: str, verbose: bool, glob: Iterable[str], components:
         A dict whose key is "datasetTypes" and whose value is a list of
         collection names.
     """
-    butler = Butler(repo, without_datastore=True)
+    butler = Butler.from_config(repo, without_datastore=True)
     expression = glob or ...
     datasetTypes = butler.registry.queryDatasetTypes(components=components, expression=expression)
     if verbose:
