+++ tests/test_packages.py
@@ -45,7 +45,7 @@ class PackagesFormatterTestCase(unittest.TestCase):
         """Create a new butler root for each test."""
         self.root = makeTestTempDir(TESTDIR)
         Butler.makeRepo(self.root)
-        self.butler = Butler(self.root, run="test_run")
+        self.butler = Butler.from_config(self.root, run="test_run")
         # No dimensions in dataset type so we don't have to worry about
         # inserting dimension data or defining data IDs.
         self.datasetType = DatasetType(
