+++ tests/test_quantumBackedButler.py
@@ -43,6 +43,7 @@ from lsst.daf.butler import (
     RegistryConfig,
     StorageClass,
 )
+from lsst.daf.butler.direct_butler import DirectButler
 from lsst.daf.butler.registry import _RegistryFactory
 from lsst.daf.butler.tests.utils import makeTestTempDir, removeTestTempDir
 from lsst.resources import ResourcePath
@@ -62,7 +63,9 @@ class QuantumBackedButlerTestCase(unittest.TestCase):
         # Make a butler and import dimension definitions.
         registryConfig = RegistryConfig(self.config.get("registry"))
         _RegistryFactory(registryConfig).create_from_config(butlerRoot=self.root)
-        self.butler = Butler(self.config, writeable=True, run="RUN")
+        butler = Butler.from_config(self.config, writeable=True, run="RUN")
+        assert isinstance(butler, DirectButler)
+        self.butler = butler
         self.butler.import_(filename=os.path.join(TESTDIR, "data", "registry", "base.yaml"))
 
         # make all dataset types
